pyproject.toml
[project]
name = "stock-agent-rag"
version = "0.1.0"
description = "Production-oriented agentic RAG stock research service with LangGraph and FastAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.57",
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.115.12",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=1.1.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.9.1",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.2.2",
    "rich>=14.0.0",
    "sec-edgar-downloader>=5.0.3",
    "sqlalchemy>=2.0.41",
    "tiktoken>=0.9.0",
    "uvicorn>=0.34.2",
    "yfinance>=0.2.66",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.8",
]

[project.scripts]
stock-agent-rag = "stock_agent_rag.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["stock_agent_rag*"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.uv]
package = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
